feat: Fetch global directory on app init
This change ensures that the global directory of users (artists, engineers, producers, stoodioz) is fetched and hydrated when the application initializes. This addresses an issue where users might be missing from various lists and views. The fetch is performed before attempting to hydrate the current user's data, guaranteeing that all user types are available from the outset.